Support basic transformations from and to the azeroth world map, as well as showing pins on it
The world map behaves slightly differently to any other zones, as
there is no global instance that covers all its "sub" zones.

To accomodate this, every of the five continents uses a "floor"
identified by its instance id, which allows the caller to indicate
which sub-continent to translate to.

Example:
Translating a pin from the center of Stormwind to the Azeroth World Map
could be done like this:

HBD:TranslateZoneCoordinates(0.5, 0.5, 301, 0, WORLDMAP_AZEROTH_ID, 0)

0.5, 0.5 are the zone coordinates in Stormwind, dead center in the map.
301 is Stormwinds map id, with floor 0 (floor nil could also be used)
WORLDMAP_AZEROTH_ID is the target zone, and the target floor is the
Eastern Kingdom instance ID (which is 0)

The Azeroth World Map supports all map functions that take a zone/level combination.
Note that in all cases the level has to be specified as the instance ID
of the origin continent of the coordinates you are trying to work with.