HBD-Pins-2.0: Hack around combat limitations in 10.1.5
SetPassThroughButtons can no longer be called in combat, but we allow
creating pins at any time during play. Until such a point when this is
fixed by Blizzard, noop out the function so that creating pins no longer
errors.

This function is called on the pin by Blizzards pin handler, which is of
course insecure on addon-created pins.